requirements.txt
temporalio>=1.7.0
docker>=7.0.0
openai>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pytest>=7.0.0
